+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 4.0.8 - 2018-06-01
 ### Added
-- Schematic now also uses the override file when exporting, fixing #48
+- Schematic now also uses the override file when exporting
+- Added support for "singles" sources
+
+### Fixed
+- Fixed issue where category site settings weren't properly imported in some cases
 
 ## 4.0.7 - 2018-05-22
 ### Fixed