+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 4.0.8 - 2018-06-01
 ### Added
-- Added support for "singles" sources, fixing #139
+- Schematic now also uses the override file when exporting
+- Added support for "singles" sources
 
 ### Fixed
 - Fixed issue where category site settings weren't properly imported in some cases